requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.1
pillow==10.1.0
python-dotenv==1.0.0
openai==1.3.7
pgvector==0.2.2
aiofiles==23.2.1
boto3==1.29.2
jinja2==3.1.2
neo4j==5.14.1
stripe==7.5.0
ffmpeg-python==0.2.0
numpy==1.26.2
opencv-python==4.8.1.78
pydub==0.25.1
email-validator==2.1.0.post1
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0
alembic==1.12.1
gunicorn==21.2.0
sentry-sdk==1.38.0
prometheus-client==0.18.0 